add common usage pattern to SDK
